计算机网络应用层-编程题

+++ 02 应用层/progs/rm_exec.sh
@@ -2,7 +2,7 @@
 
 for file in `find . -type f -print`
 do
-    if [ -x $file ]
+    if [ -x $file ] && [[ "`basename $file`" != "`basename $0`" ]] 
     then
         rm $file
     fi
